+++ express/code/blocks/frictionless-quick-action/frictionless-quick-action.js
@@ -2,7 +2,7 @@ import { transformLinkToAnimation } from '../../scripts/utils/media.js';
 import { getLibs, getIconElementDeprecated, decorateButtonsDeprecated } from '../../scripts/utils.js';
 import { buildFreePlanWidget } from '../../scripts/widgets/free-plan.js';
 import { sendFrictionlessEventToAdobeAnaltics } from '../../scripts/instrument.js';
-import { createLocaleDropdownWrapper } from '../../scripts/widgets/locale-dropdown.js';
+import { createLocaleDropdownWrapper } from '../../scripts/widgets/frictionless-locale-dropdown.js';
 
 let createTag; let getConfig;
 let getMetadata;
